+++ Sthor/Sources/Models/Ui.swift
@@ -11,11 +11,11 @@ import SwiftUI
 struct Header: Identifiable {
     let id = UUID()
     let title: LocalizedStringKey
-    let subTitle: LocalizedStringKey?
+    let subTitle: String?
     let horizontalAlignment: Alignment
     let verticalAlignment: Alignment
     
-    init(title: LocalizedStringKey, subTitle: LocalizedStringKey? = nil, horizontalAlignment: Alignment = .center, verticalAlignment: Alignment = .center) {
+    init(title: LocalizedStringKey, subTitle: String? = nil, horizontalAlignment: Alignment = .center, verticalAlignment: Alignment = .center) {
         self.title = title
         self.subTitle = subTitle
         self.horizontalAlignment = horizontalAlignment