+++ Sthor/Sources/Utils/Date+Extensions.swift
@@ -1,0 +1,15 @@
+//
+//  Date+Extensions.swift
+//  Sthor
+//
+//  Created by Anh Cao on 5.5.2024.
+//
+
+import SwiftUI
+
+extension Date {
+    var currentDate: Date {
+        return Date()
+    }
+    
+}