+++ Sthor/Sources/Utils/View+Extenstions.swift
@@ -19,4 +19,10 @@ extension View {
         self
             .frame(maxHeight: .infinity, alignment: alignment)
     }
+    
+    func formatDate(date: Date, format: String? = "EEEE, d'th' MMMM yyyy") -> String {
+        let formatter = DateFormatter()
+        formatter.dateFormat = format
+        return formatter.string(from: date)
+    }
 }