+++ CHANGELOG.md
@@ -7,4 +7,4 @@ How the application has been growth so far?
 - Start the project 1st May 2024 
 - Create folder structure 
 - Create 4 basic templates for tab views: Home, Electric, Profile and Notifications
-
+- Dynamic get current date